RDM-5545 Add unit tests around true, false and null values of ignoreWarning flag for CallbackService. Rename tests that use wiremock accordingly. Small naming cleanups.
